added cli util to build connections to project directories. added art to home page
