+++ README.md
@@ -3,7 +3,10 @@
 ## About
 Welcome to my coding portfolio! Here I have setup several coding projects that I built in addition to some of my other hobbies like photography and videography
 
+# Release Schedule
+I intend to update this portfolio with two main releases while occasionally making small updates.
+
 ## Built With
 - React
 - Firebase
-
+- p5
