+++ client/.gitignore
@@ -1,5 +1,8 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
+# project files
+/src/utils/projectRoutes.js
+
 # dependencies
 /node_modules
 /.pnp
