+++ client/README.md
@@ -1,0 +1,4 @@
+# Portfolio
+
+## Summary
+A website to display my projects and some info about myself
