+++ client/index.html
@@ -1,0 +1,54 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <link rel="icon" href="/portfolioicon_6UN_icon.ico" />
+    <meta
+      name="viewport"
+      content="width=device-width, initial-scale=1.0, minimum-scale=1.0"
+    />
+    <meta name="theme-color" content="#000000" />
+    <meta
+      name="description"
+      ,
+      minimum-scale="1.0"
+      content="Web Developer, Photographer, Videographer"
+    />
+
+    <link rel="apple-touch-icon" href="logo192.png" />
+    <!--
+      manifest.json provides metadata used when your web app is installed on a
+      user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
+    -->
+    <link rel="manifest" href="/manifest.json" />
+
+    <!-- <link
+      href="https://fonts.googleapis.com/css?family=Lato&display=swap"
+      rel="stylesheet"
+    /> -->
+    <!-- 
+    <script
+      src="https://kit.fontawesome.com/5bdec6d93a.js"
+      crossorigin="anonymous"
+    ></script> -->
+
+    <!-- should i bring this in -->
+    <link rel="stylesheet" href="reset.css" />
+    <!--
+      Notice the use of %PUBLIC_URL% in the tags above.
+      It will be replaced with the URL of the `public` folder during the build.
+      Only files inside the `public` folder can be referenced from the HTML.
+
+      Unlike "/favicon.ico" or "favicon.ico", "%PUBLIC_URL%/favicon.ico" will
+      work correctly both with client-side routing and a non-root public URL.
+      Learn how to configure a non-root public URL by running `npm run build`.
+    -->
+    <title>Daniel Chicchon</title>
+  </head>
+
+  <body>
+    <noscript>Web Developer, Photographer, Videographer</noscript>
+    <div id="root"></div>
+    <script type="module" src="./src/index.jsx"></script>
+  </body>
+</html>
