+++ client/package.json
@@ -2,28 +2,33 @@
   "name": "client",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "dependencies": {
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.10",
+    "chai": "^4.3.7",
     "firebase": "^9.4.0",
-    "mapbox-gl": "^2.6.0",
+    "gh-pages": "^4.0.0",
+    "mapbox-gl": "^2.12.0",
+    "mocha": "^10.2.0",
     "p5": "^1.5.0",
     "react": "^17.0.2",
     "react-color": "^2.19.3",
+    "react-dnd": "^16.0.1",
+    "react-dnd-html5-backend": "^16.0.1",
     "react-dom": "^17.0.2",
     "react-router-dom": "^6.0.2",
-<<<<<<< HEAD
     "react-scripts": "^4.0.3",
-    "three": "^0.137.0"
-=======
-    "react-scripts": "^4.0.3"
->>>>>>> cbecaf8 (remove unused packages)
+    "web-vitals": "^3.1.1",
+    "zustand": "^4.3.2"
   },
   "scripts": {
-    "start": "set PORT=3005 && react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "start": "vite --host",
+    "loadProjects": "node ./utils",
+    "build": "vite build",
+    "serve": "vite preview",
     "update-deps": "npx npm-check-updates -u && npm install",
-    "deploy": "react-scripts build && firebase deploy"
+    "deploy": "vite build && firebase deploy"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -40,6 +45,11 @@
     ]
   },
   "devDependencies": {
-    "serve": "^13.0.2"
+    "@vitejs/plugin-react": "^3.0.1",
+    "chalk": "^5.2.0",
+    "serve": "^13.0.2",
+    "vite": "^4.0.4",
+    "vite-plugin-svgr": "^2.4.0",
+    "vite-tsconfig-paths": "^4.0.3"
   }
 }
