+++ client/src/App.css
@@ -1,23 +1,40 @@
-.App {
+body {
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen", "Ubuntu",
+    "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+html,
+body {
+  background: rgb(31, 33, 34);
   font-family: "Lato", sans-serif;
+  height: 100%;
 }
 
-/* width */
-/* ::-webkit-scrollbar {
-  width: 7px;
-} */
+h2 {
+  font-size: large;
+}
 
-/* Track */
-/* ::-webkit-scrollbar-track {
-  display: none;
-} */
+#root {
+  height: 100%;
+}
 
-/* Handle */
-/* ::-webkit-scrollbar-thumb {
-  background: rgb(173, 170, 170);
-} */
+a {
+  color: white;
+  text-decoration: none;
+}
 
-/* Handle on hover */
-/* ::-webkit-scrollbar-thumb:hover {
-  background: #555;
-} */
+a:hover {
+  text-decoration: underline;
+}
+
+/* Helper CSS Pointers */
+
+.flex-end {
+  justify-content: flex-end;
+}
+
+.flex-start {
+  justify-content: flex-start;
+}
