--- client/src/Components/FadeIn/style.css
@@ -1,28 +1,0 @@
-.fade-in-section {
-    opacity: 0;
-    transform: translateY(5vh);
-    visibility: hidden;
-    /* transition: opacity 0.8s ease-out; */
-    transition: opacity 0.8s ease-out, transform 1s ease-out;
-    will-change: opacity, visibility;
-}
-
-
-.fade-in-section.is-visible {
-    /* animation-name: fillIn; */
-    /* animation-duration: 1s; */
-    /* animation-fill-mode: forwards; */
-    transform: none;
-    opacity: 1;
-    visibility: visible;
-}
-
-/* @keyframes fillIn { */
-    /* 0% { */
-        /* opacity: 0; */
-    /* } */
-    
-    /* 100% { */
-        /* opacity: 1; */
-    /* } */
-/* } */
