--- client/src/Components/Footer/style.css
@@ -1,30 +1,0 @@
-#footer {
-  color: white;
-  padding: 5rem 0 1.5rem;
-  position: relative;
-  z-index: 1;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  /* background-color: ; */
-  /* background: linear-gradient(to top, hsl(210, 60%, 40%), hsl(210, 40%, 50%)); */
-}
-
-.footer-sub {
-  opacity: 0.8;
-  font-size: 0.75rem;
-  text-align: center;
-  position: relative;
-}
-
-@media only screen and (max-width: 414px) {
-  #footer {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-    height: 4rem;
-    /* bottom: 0; */
-  }
-}
