--- client/src/Components/Navbar/style.css
@@ -1,50 +1,0 @@
-.nav-show {
-  position: fixed;
-  width: 100%;
-  height: 75px;
-  top: 0;
-  background: rgb(36, 35, 35);
-  transition: top 0.6s;
-  z-index: 100;
-}
-
-.nav-area {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  text-align: center;
-  height: 100%;
-  padding: 0 1rem;
-}
-
-.nav-title {
-  font-size: 3rem;
-}
-
-.nav-links ul {
-  list-style-type: none;
-  padding: 0;
-  margin: 0;
-  display: flex;
-}
-
-.spacer {
-  flex: 1;
-}
-
-.nav-links ul li a {
-  float: left;
-  padding: 0 0.5rem;
-}
-
-@media only screen and (max-width: 555px) {
-  .nav-show {
-    position: fixed;
-    width: 100%;
-    height: 50px;
-    top: 0;
-    background: rgb(36, 35, 35);
-    transition: top 0.6s;
-    z-index: 100;
-  }
-}
