--- client/src/Constants/routes.js
@@ -1,13 +1,0 @@
-export const HOME = "/";
-export const CODE = "/code";
-
-// Code
-export const POLUS = "/polus";
-
-// Polus
-export const POLUS_ABOUT = "/polus/about";
-export const POLUS_PRIVACY = "/polus/privacy";
-export const POLUS_TERMS = "/polus/terms";
-
-// Topo
-export const TOPO = '/topo'
