--- client/src/Pages/Home/style.css
@@ -1,7 +1,0 @@
-.main {
-    padding: 0.1rem 1rem;
-    color: white;
-    /* height: 100vh; */
-    /* width: 100vw; */
-    background: rgb(31, 33, 34);
-}
