+++ client/src/Utils/mainRoutes.js
@@ -1,0 +1,5 @@
+// DEFAULT
+export const HOME = { to: "/", text: "Home" };
+export const PROJECTS = { to: "/projects", text: "Projects" };
+export const ABOUT = { to: "/about", text: "About" };
+export const ERROR = { to: '*', text: "Error" };
