+++ client/src/Utils/routes.js
@@ -1,0 +1,16 @@
+export const HOME = { to: "/", text: "Home" };
+export const PROJECTS = { to: "/projects", text: "Projects" };
+export const ABOUT = { to: "/about", text: "About" };
+export const ERROR = { to: '*', text: "Error" };
+
+// Polus
+export const POLUS = { to: "/polus", text: 'Polus' };
+export const POLUS_ABOUT = { to: "/polus/about", text: 'About' };
+export const POLUS_PRIVACY = { to: "/polus/privacy", text: 'Privacy' };
+export const POLUS_TERMS = { to: "/polus/terms", text: 'Terms' };
+
+// Topo
+// export const TOPO = { to: '/topo', text: 'Topo' }
+
+// LineGrid
+// export const LINEGRID = { to: '/linegrid', text: 'LineGrid' }
