+++ client/src/projects/Polus/pages/Terms/Terms.jsx
@@ -1,0 +1,287 @@
+import React from "react";
+import { Link } from "react-router-dom";
+
+import Navbar from "../../components/Navbar";
+import Footer from "../../components/Footer";
+import "./Terms.css";
+
+function Terms() {
+  return (
+    <>
+      <Navbar />
+      <nav className="legal-nav">
+        <Link to="/polus/terms">Terms of Use</Link>
+        <Link to="/polus/privacy">Privacy Policy</Link>
+      </nav>
+      <section className="legal legal-terms">
+        <h2>Terms of Use</h2>
+        <p className="small">
+          Important: Before using this software (Polus), please carefully read
+          this agreement which contains the terms and conditions under which you
+          are acquiring a license to use Polus. If you do not accept the terms
+          and conditions of this agreement please do not use Polus. If you
+          access or use Polus, you will be accepting the terms and conditions of
+          this agreement and Polus' Privacy Policy. Polus is protected by
+          copyright laws and international copyright treaties, as well as other
+          intellectual property laws.
+        </p>
+        <p className="small">
+          If you are an agent or employee of another entity you represent and
+          warrant that (i) the individual accepting this agreement is duly
+          authorized to accept this agreement on such entity’s behalf and to
+          bind such entity, and (ii) such entity has full power, corporate or
+          otherwise, to enter into this agreement and perform its obligations
+          hereunder.
+        </p>
+        <h4 id="1-definitions">1. Definitions</h4>
+        <p>
+          (i) "User" or "You" means the individual or the business entity to
+          which a license is granted by Polus for the use of the Software;{" "}
+        </p>
+        <p>
+          (ii) "Effective Date" means the date on which a User first installs
+          the Licensed Software through Polus' website or through a third-party
+          application, extension, or add-on store such as the Chrome Web Store
+        </p>
+        <p>
+          (iii) “Licensed Software” means the web-based software extension
+          called Polus, which is proprietary to Polus;
+        </p>
+        <p>
+          (iv) “Licensee Data” means all electronic data or information
+          submitted by User to the Service;
+        </p>
+        <p>
+          (v) “Service” means the provision of access to the Licensed Software
+          for the purpose of transmitting Licensee Data;
+        </p>
+        <p>
+          (vi) “Usage Data” means data collected about the User’s use of the
+          Service. For example, how often the User accesses a to do list, or
+          what photo(s) they download.
+        </p>
+        <h4 id="2-licenses-and-data">2. Licenses and Data</h4>
+        <p>
+          <strong> 2. Licenses and Data 2.1 Software License Grant</strong>
+          <br />
+          Subject to all of the terms and conditions of this Agreement and
+          payment of any applicable fees, Polus grants to the User a
+          royalty-free, non-exclusive, non-transferable license to download,
+          install and use the Licensed Software as part of the Service.
+        </p>
+        <p>
+          <strong> 2.2 License to Licensee Data</strong>
+          <br />
+          User grants to Polus a non-exclusive license to access and modify
+          Licensee Data as required in order to provide the Service.
+        </p>
+        <p>
+          <strong> 2.3 Usage Data</strong>
+          <br />
+          User agrees that Polus will have the right to collect Usage Data and
+          to create statistics and analytics therefrom (“Derived Data”) as per
+          the Privacy Policy currently in effect.
+        </p>
+        <p>
+          <strong> 2.4 Consent for Children Under 16</strong>
+          <br />
+          By using our service, you represent and warrant that you are at least
+          16 years of age. No one under the age of 16 may use the Service,
+          except as set forth as follows.
+        </p>
+        <p>
+          Polus is not directed to children, and we expect that any use by
+          children will only be done with the guidance, supervision and consent
+          of their parents, guardians and/or authorized school officials. If you
+          are under the age of 16 you must receive consent from a parent,
+          guardian and/or an authorized school official to use the Service.
+          Polus relies on parents and guardians to ensure that minors only use
+          Polus if they can understand their rights and responsibilities as laid
+          out in these Terms of Service and in our Privacy Policy.
+        </p>
+        <p>
+          If you are an authorized school official in the United States or in
+          the EU and want your students, who are under the age of 16, to create
+          Polus accounts, you are responsible for complying with the U.S.
+          Children's Online Privacy Protection Act ("COPPA") and, to the extent
+          applicable, the Family Educational Rights and Privacy Act ("FERPA"),
+          or the GDPR policy, respectively. This means that you must notify the
+          students' parents/guardians of the personally identifiable information
+          that Polus will collect, and that you will obtain parental/guardian
+          consent before your students establish accounts or use Polus. When
+          obtaining such consent, you must provide their parents or guardians
+          with a copy of our Privacy Policy and keep all consents on file and
+          provide them to us upon request.
+        </p>
+        <p>
+          For more information on complying with COPPA, see the Federal Trade
+          Commission's website. If you are located outside of the United States,
+          we will rely on you to obtain any required consent or approval from
+          the parent or guardian of any student covered by similar laws, and, as
+          a condition to your and your students' use of Polus, you agree that
+          you will be responsible for complying with such laws.
+        </p>
+        <p>
+          If we become aware that a child under the age of 16 has provided us
+          with personal information, without the consent of a parent, guardian
+          and/or an authorized school official, we will delete the child’s
+          account and information. If you are aware that a child under the age
+          of 16 has provided Polus with personal information without the consent
+          of a parent, guardian and/or an authorized school official, please
+          contact Polus support.
+        </p>
+        <h4 id="3-disclaimer-of-warranties"> 3. Disclaimer of Warranties </h4>
+        <p>
+          <strong>3.1</strong> To the greatest extent permitted by law, the
+          licensed software and technical support provided by Polus hereunder
+          are provided on an "as is" basis. There are no warranties,
+          representations or conditions, express or implied, written or oral,
+          arising by statute, operation of law, course of dealing, usage of
+          trade or otherwise, regarding them or any other product or service
+          provided under this agreement or in connection with this agreement by
+          Polus. Polus disclaims any implied warranties or conditions of
+          quality, merchantability, merchantable quality, durability, fitness
+          for a particular purpose and non-infringement. Polus does not
+          represent or warrant that the software shall meet any or all of User’s
+          particular requirements, that the software will operate error-free or
+          uninterrupted or that all errors or defects in the software can be
+          found or corrected.
+        </p>
+        <p>
+          <strong>3.2</strong> In certain jurisdictions some or all of the
+          provisions in this Section may not be effective, or the applicable law
+          may mandate a more extensive warranty, in which case the applicable
+          law will prevail over this Agreement.
+        </p>
+        <h4 id="4-limitation-of-liability"> 4. Limitation of Liability </h4>
+        <p>
+          <strong>4.1</strong> To the greatest extent permitted by applicable
+          law, in no event shall Polus be liable to the User or any other person
+          for any direct, indirect, incidental, special, punitive, exemplary or
+          consequential damages whatsoever, including without limitation, legal
+          expenses, loss of business, loss of profits, loss of revenue, lost or
+          damaged data, loss of computer time, cost of substitute goods or
+          services, or failure to realize expected savings or any other
+          commercial or economic losses arising out of or in connection with
+          this agreement, even if Polus has been advised of the possibility of
+          such loss or damages, or such losses or damages are foreseeable.
+        </p>
+        <p>
+          <strong>4.2</strong> The entire liability of Polus and the User’s
+          exclusive remedy with respect to the software and technical support.
+          Any other products or services supplied by Polus in connection with
+          this agreement for damages for any cause and regardless of the cause
+          of action, whether in contract or in tort, including fundamental
+          breach or negligence, will be limited in the aggregate to the amounts
+          paid by User to Polus under this agreement for the software, technical
+          support or services giving rise to the claim.
+        </p>
+        <p>
+          <strong>4.3</strong> The disclaimer of representations, warranties and
+          conditions and limitation of liability constitute an essential part of
+          this agreement. You acknowledge that for the disclaimer of
+          representations, warranties and conditions and limitation of
+          liability, neither Polus nor any of its licensors or suppliers would
+          grant the rights granted in this agreement.
+        </p>
+        <h4 id="5-proprietary-rights">5. Proprietary Rights </h4>
+        <p>
+          <strong>5.1</strong> The User acknowledges and agrees that the
+          Licensed Software contains confidential and proprietary information
+          and trade secrets belonging to Polus and its licensors. The User
+          acknowledges and agrees that title in and rights to the Software
+          remains exclusively with Polus and its licensors. The User’s rights to
+          the Software are strictly limited to those granted in this Agreement.
+          User shall not decompile, disassemble or otherwise reverse engineer
+          the Software. If the foregoing provision is prohibited by applicable
+          law, the User will provide Polus with advance written notification of
+          (a) its intention to decompile, disassemble or otherwise reverse
+          engineer the Software, and (b) the nature of the work involved. Polus
+          will be given the right of first refusal to perform such work at its
+          prevailing rates and prices.
+        </p>
+        <h4 id="6-uninstalling-the-software"> 6. Uninstalling the Software </h4>
+        <p>
+          <strong>6.1</strong> Users can uninstall the Licensed Software at any
+          time by following the process outlined in your browser’s help
+          documentation.
+        </p>
+        <h4 id="7-term-and-termination">7. Term and Termination </h4>
+        <p>
+          <strong>7.1</strong> The term of this Agreement will begin upon the
+          Effective Date and shall continue for as long as User uses the Service
+          under subscription and pays all applicable fees, unless earlier
+          terminated under this section 7.
+        </p>
+        <p>
+          <strong>7.2</strong> Polus may terminate this Agreement in the event
+          of any breach by the User if such breach has not been cured within
+          thirty (30) days of notice to User. No termination of this Agreement
+          will entitle the User to a refund of any amounts paid by User to Polus
+          or affect any obligations User may have to pay any outstanding amounts
+          owing to Polus.
+        </p>
+        <p>
+          <strong>7.3</strong> The User’s rights to use and access the Service
+          will immediately terminate upon termination or expiration of this
+          Agreement. Sections 1, 4, 5, 6, 7, and 8.3 of this Agreement shall
+          survive the expiration or termination of this Agreement.
+        </p>
+        <h4 id="8-general-provisions">8. General Provisions </h4>
+        <p>
+          <strong>8.1 No Waiver</strong>
+          <br />
+          No delay or failure in exercising any right under this Agreement, or
+          any partial or single exercise of any right, will constitute a waiver
+          of that right or any other rights under this Agreement. No consent to
+          a breach of any express or implied term set out in this Agreement
+          constitutes consent to any subsequent breach, whether of the same or
+          any other provision.
+        </p>
+        <p>
+          <strong>8.2 Severability</strong>
+          <br />
+          If any provision of this Agreement is, or becomes, unenforceable, it
+          will be severed from this Agreement and the remainder of this
+          Agreement will remain in full force and effect.
+        </p>
+        <p>
+          <strong>8.3 Assignment</strong>
+          <br />
+          The User may not transfer or assign this Agreement (whether
+          voluntarily, by operation of law, or otherwise) without Polus'
+          prior written consent. Polus may assign this Agreement at any time
+          without notice. This Agreement is binding upon and will inure to the
+          benefit of both parties, and their respective successors and permitted
+          assigns.
+        </p>
+        <p>
+          <strong>8.4 Governing Law and Venue</strong>
+          <br />
+          This Agreement shall be governed by the laws of the United States of
+          America. No choice of laws rules of any jurisdiction shall apply to
+          this Agreement. The User consents and agrees that the courts of the
+          United States of America shall have jurisdiction over any legal action
+          or proceeding brought by the User arising out of or relating to this
+          Agreement, and the User consents to the jurisdiction of such courts
+          for any such action or proceeding.
+        </p>
+        <p>
+          <strong>8.5 Entire Agreement </strong>
+          <br />
+          This Agreement is the entire understanding and agreement between the
+          User and Polus with respect to the subject matter hereof, and it
+          supersedes all prior negotiations, commitments and understandings,
+          verbal or written, any purchase order issued by User.
+        </p>
+        <p>
+          If you have any questions about this agreement, please contact us.
+        </p>
+        <p className="last-updated">Last updated September 14, 2020.</p>
+      </section>
+      <Footer />
+    </>
+  );
+}
+
+export default Terms;
